# docs/swagger-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lumina AI API",
    "description": "Lumina AI 图片处理应用后端API文档\n\n本API提供以下核心功能：\n- 用户认证与授权\n- AI问答与推荐\n- 图片上传与处理（单图/批量）\n- 作品管理\n- 订阅与支付\n- 场景配置管理\n\n认证方式：使用Bearer Token认证，在Header中添加 `Authorization: Bearer {token}`",
    "version": "1.0.0",
    "contact": {
      "name": "Lumina AI API Support",
      "email": "api-support@lumina.ai"
    }
  },
  "servers": [
    {
      "url": "https://api.lumina.ai/v1",
      "description": "生产环境"
    },
    {
      "url": "https://api-staging.lumina.ai/v1",
      "description": "测试环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/user/profile": {
      "get": {
        "tags": ["User"],
        "summary": "获取用户信息",
        "description": "获取当前登录用户的详细信息",
        "operationId": "getUserProfile",
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["User"],
        "summary": "更新用户信息",
        "description": "更新当前用户的基本信息",
        "operationId": "updateUserProfile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/user/stats": {
      "get": {
        "tags": ["User"],
        "summary": "获取用户统计信息",
        "description": "获取用户的处理统计、剩余配额等信息",
        "operationId": "getUserStats",
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStats"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/works": {
      "get": {
        "tags": ["Works"],
        "summary": "获取作品列表",
        "description": "获取用户的所有作品，支持分页和筛选",
        "operationId": "getWorks",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码（从1开始）",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "筛选分类",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["all", "taobao", "douyin", "xiaohongshu", "amazon", "custom"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorksListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Works"],
        "summary": "保存作品",
        "description": "将处理后的图片保存为作品",
        "operationId": "saveWork",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveWorkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "保存成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Work"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/works/{workId}": {
      "get": {
        "tags": ["Works"],
        "summary": "获取作品详情",
        "description": "获取指定作品的详细信息",
        "operationId": "getWorkDetail",
        "parameters": [
          {
            "name": "workId",
            "in": "path",
            "required": true,
            "description": "作品ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkDetail"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Works"],
        "summary": "删除作品",
        "description": "删除指定的作品",
        "operationId": "deleteWork",
        "parameters": [
          {
            "name": "workId",
            "in": "path",
            "required": true,
            "description": "作品ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/works/batch-delete": {
      "post": {
        "tags": ["Works"],
        "summary": "批量删除作品",
        "description": "批量删除多个作品",
        "operationId": "batchDeleteWorks",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/subscription/current": {
      "get": {
        "tags": ["Subscription"],
        "summary": "获取当前订阅",
        "description": "获取当前用户的订阅信息",
        "operationId": "getCurrentSubscription",
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentSubscriptionResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "用户未订阅",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/subscription/create-order": {
      "post": {
        "tags": ["Subscription"],
        "summary": "创建订阅订单",
        "description": "创建订阅订单并返回支付信息",
        "operationId": "createSubscriptionOrder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "订单创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/subscription/payment-callback": {
      "post": {
        "tags": ["Subscription"],
        "summary": "支付回调",
        "description": "支付平台回调接口（微信支付/支付宝）",
        "operationId": "paymentCallback",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCallbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "回调处理成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/settings": {
      "get": {
        "tags": ["Settings"],
        "summary": "获取应用设置",
        "description": "获取用户的应用设置（通知、自动保存等）",
        "operationId": "getSettings",
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettings"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["Settings"],
        "summary": "更新应用设置",
        "description": "更新用户的应用设置",
        "operationId": "updateSettings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettings"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "使用JWT Token进行认证，格式：Bearer {token}"
      }
    },
    "schemas": {
...
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "user_123456"
          },
          "phoneNumber": {
            "type": "string",
            "example": "138****5678"
          },
          "nickname": {
            "type": "string",
            "example": "创作者"
          },
          "avatar": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "example": "https://cdn.lumina.ai/avatars/user_123456.jpg"
          },
          "isPro": {
            "type": "boolean",
            "description": "是否为专业版会员",
            "example": true
          },
          "membershipType": {
            "type": "string",
            "enum": ["free", "monthly", "annual", "enterprise"],
            "example": "annual"
          },
          "membershipExpiry": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2025-01-28T00:00:00Z"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00Z"
          }
        }
      },
      "UpdateUserProfileRequest": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20,
            "example": "新昵称"
          },
          "avatar": {
            "type": "string",
            "format": "uri",
            "example": "https://cdn.lumina.ai/avatars/new_avatar.jpg"
          }
        }
      },
      "UserStats": {
        "type": "object",
        "properties": {
          "processedCount": {
            "type": "integer",
            "description": "已处理图片数量",
            "example": 1234
          },
          "remainingQuota": {
            "type": "integer",
            "description": "剩余配额（-1表示无限）",
            "example": 58
          },
          "dailyQuota": {
            "type": "integer",
            "description": "每日配额（-1表示无限）",
            "example": -1
          },
          "membershipDaysLeft": {
            "type": "integer",
            "nullable": true,
            "description": "会员剩余天数",
            "example": 28
          },
          "storageUsed": {
            "type": "number",
            "description": "已使用存储空间（GB）",
            "example": 2.3
          },
          "storageTotal": {
            "type": "number",
            "description": "总存储空间（GB，-1表示无限）",
            "example": 5.0
          }
        }
      },
...
      "UploadedImage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "img_abc123"
          },
          "filename": {
            "type": "string",
            "example": "IMG_2024.jpg"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://cdn.lumina.ai/uploads/img_abc123.jpg"
          },
          "thumbnail": {
            "type": "string",
            "format": "uri",
            "example": "https://cdn.lumina.ai/uploads/thumb_img_abc123.jpg"
          },
          "width": {
            "type": "integer",
            "example": 1920
          },
          "height": {
            "type": "integer",
            "example": 1080
          },
          "size": {
            "type": "integer",
            "description": "文件大小（字节）",
            "example": 2048000
          },
          "format": {
            "type": "string",
            "enum": ["jpg", "png", "webp"],
            "example": "jpg"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
...
      "ImageOperation": {
        "type": "object",
        "required": ["type"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["cutout", "background", "lighting", "filter", "resize"],
            "description": "操作类型"
          },
          "params": {
            "type": "object",
            "description": "操作参数（根据type不同而不同）",
            "additionalProperties": true,
            "example": {
              "backgroundColor": "#FFFFFF",
              "backgroundTemplateId": "white",
              "brightness": 1.2,
              "contrast": 1.1
            }
          }
        }
      },
...
      "ProcessedImage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "processed_img_xyz789"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://cdn.lumina.ai/processed/processed_img_xyz789.jpg"
          },
          "thumbnail": {
            "type": "string",
            "format": "uri",
            "example": "https://cdn.lumina.ai/processed/thumb_processed_img_xyz789.jpg"
          },
          "width": {
            "type": "integer",
            "example": 2000
          },
          "height": {
            "type": "integer",
            "example": 2000
          },
          "size": {
            "type": "integer",
            "description": "文件大小（字节）",
            "example": 1536000
          },
          "format": {
            "type": "string",
            "enum": ["jpg", "png", "webp"],
            "example": "jpg"
          },
          "operations": {
            "type": "array",
            "description": "应用的处理操作",
            "items": {
              "$ref": "#/components/schemas/ImageOperation"
            }
          }
        }
      },
      "SaveWorkRequest": {
        "type": "object",
        "required": ["processedImageId", "filename"],
        "properties": {
          "processedImageId": {
            "type": "string",
            "description": "处理后的图片ID",
            "example": "processed_img_xyz789"
          },
          "filename": {
            "type": "string",
            "description": "作品文件名",
            "example": "IMG_2024_processed.jpg"
          },
          "category": {
            "type": "string",
            "enum": ["taobao", "douyin", "xiaohongshu", "amazon", "custom"],
            "description": "作品分类",
            "example": "taobao"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "标签",
            "example": ["产品图", "白底图"]
          }
        }
      },
      "Work": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "work_123456"
          },
          "filename": {
            "type": "string",
            "example": "IMG_2024_processed.jpg"
          },
          "thumbnail": {
            "type": "string",
            "format": "uri",
            "example": "https://cdn.lumina.ai/works/thumb_work_123456.jpg"
          },
          "category": {
            "type": "string",
            "enum": ["taobao", "douyin", "xiaohongshu", "amazon", "custom"],
            "example": "taobao"
          },
          "size": {
            "type": "integer",
            "description": "文件大小（字节）",
            "example": 1536000
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
      "WorkDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Work"
          },
          {
            "type": "object",
            "properties": {
              "imageUrl": {
                "type": "string",
                "format": "uri",
                "example": "https://cdn.lumina.ai/works/work_123456.jpg"
              },
              "beforeImage": {
                "$ref": "#/components/schemas/UploadedImage"
              },
              "afterImage": {
                "$ref": "#/components/schemas/ProcessedImage"
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": ["产品图", "白底图"]
              },
              "operations": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ImageOperation"
                }
              }
            }
          }
        ]
      },
      "WorksListResponse": {
        "type": "object",
        "properties": {
          "works": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Work"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "totalStorage": {
            "type": "number",
            "description": "总存储使用量（字节）",
            "example": 2457600000
          }
        }
      },
      "BatchDeleteRequest": {
        "type": "object",
        "required": ["workIds"],
        "properties": {
          "workIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "example": ["work_123456", "work_789012"]
          }
        }
      },
...
      "CurrentSubscriptionResponse": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "string",
            "example": "annual"
          },
          "planName": {
            "type": "string",
            "example": "年度会员"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00Z"
          },
          "expiryDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-01T00:00:00Z"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "autoRenew": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "CreateOrderRequest": {
        "type": "object",
        "required": ["planId"],
        "properties": {
          "planId": {
            "type": "string",
            "enum": ["monthly", "annual"],
            "example": "annual"
          },
          "paymentMethod": {
            "type": "string",
            "enum": ["wechat", "alipay"],
            "default": "wechat",
            "example": "wechat"
          }
        }
      },
      "OrderResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "example": "order_abc123"
          },
          "amount": {
            "type": "number",
            "description": "订单金额（元）",
            "example": 299
          },
          "paymentInfo": {
            "type": "object",
            "description": "支付信息（根据支付方式不同）",
            "additionalProperties": true,
            "example": {
              "qrCode": "https://api.lumina.ai/payment/qr/order_abc123",
              "paymentUrl": "weixin://wxpay/bizpayurl?pr=xxx"
            }
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "订单过期时间",
            "example": "2024-01-15T11:00:00Z"
          }
        }
      },
      "PaymentCallbackRequest": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "example": "order_abc123"
          },
          "paymentMethod": {
            "type": "string",
            "enum": ["wechat", "alipay"],
            "example": "wechat"
          },
          "transactionId": {
            "type": "string",
            "example": "wx_transaction_123456"
          },
          "amount": {
            "type": "number",
            "example": 299
          },
          "signature": {
            "type": "string",
            "description": "支付平台签名",
            "example": "signature_string"
          },
          "timestamp": {
            "type": "integer",
            "example": 1705315200
          }
        }
      },
...
      "AppSettings": {
        "type": "object",
        "properties": {
          "notifications": {
            "type": "boolean",
            "description": "是否启用通知",
            "example": true
          },
          "autoSave": {
            "type": "boolean",
            "description": "是否自动保存",
            "example": true
          },
          "defaultQuality": {
            "type": "integer",
            "minimum": 60,
            "maximum": 100,
            "description": "默认图片质量",
            "example": 85
          },
          "defaultFormat": {
            "type": "string",
            "enum": ["jpg", "png", "webp"],
            "description": "默认图片格式",
            "example": "jpg"
          }
        }
      },
      "UpdateSettingsRequest": {
        "type": "object",
        "properties": {
          "notifications": {
            "type": "boolean"
          },
          "autoSave": {
            "type": "boolean"
          },
          "defaultQuality": {
            "type": "integer",
            "minimum": 60,
            "maximum": 100
          },
          "defaultFormat": {
            "type": "string",
            "enum": ["jpg", "png", "webp"]
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "pageSize": {
            "type": "integer",
            "example": 20
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "INVALID_REQUEST"
              },
              "message": {
                "type": "string",
                "example": "请求参数错误"
              },
              "details": {
                "type": "object",
                "additionalProperties": true,
                "description": "错误详情"
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未授权，需要登录",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}